requirements.txt
<<<<<<< HEAD
Django	5.0.4	5.0.4
asgiref	3.8.1	3.8.1
django-jazzmin	3.0.0	3.0.0
djangorestframework	3.15.1	3.15.1
pillow	10.3.0	10.3.0
pip	21.3.1	24.0
python-decouple	3.8	3.8
setuptools	60.2.0	69.5.1
sqlparse	0.5.0	0.5.0
typing-extensions	4.11.0	4.11.0
tzdata	2024.1	2024.1
wheel	0.37.1	0.43.0
=======
asgiref==3.8.1
black==24.4.2
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
dicttoxml==1.7.16
Django==5.0.4
django-ckeditor==6.7.1
django-ckeditor-5==0.2.12
django-jazzmin==3.0.0
django-jet==1.0.8
django-js-asset==2.2.0
django-phonenumber-field==7.3.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
firebase-admin==6.5.0
google-api-core==2.19.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
jet==1.0
msgpack==1.0.8
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
phonenumbers==8.13.35
pillow==10.3.0
platformdirs==4.2.1
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
python-decouple==3.8
requests==2.31.0
rsa==4.9
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
xmltodict==0.13.0
>>>>>>> 8988b8e3826159e9674ca5026bda2db933febf1e
